requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.35
pymysql==1.1.1
python-multipart==0.0.9
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python.dotenv==1.0.1
pydantic[email]==2.9.2